+++ src/main/java/duke/exception/DukeException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exception;
 
 public class DukeException extends RuntimeException {
 
@@ -12,73 +12,5 @@ public class DukeException extends RuntimeException {
     }
 }
 
-class InvalidCommand extends DukeException {
 
-    private final String errorMessage;
-    /**
-     * Initialises a new instance of InvalidCommand.
-     *
-     * @param errorMessage The error message that is related to the exception.
-     */
-    public InvalidCommand(String errorMessage) {
-        super(errorMessage);
-        this.errorMessage = errorMessage;
-    }
 
-    /**
-     * Creates an error message to be displayed to the user.
-     *
-     * @return String containing the output message.
-     */
-    @Override
-    public String toString() {
-        return "OOPS!!! " + this.errorMessage;
-    }
-}
-
-class InvalidIndex extends DukeException {
-
-    private final String errorMessage;
-    /**
-     * Initialises a new instance of InvalidIndex.
-     *
-     * @param errorMessage The error message that is related to the exception.
-     */
-    public InvalidIndex(String errorMessage) {
-        super(errorMessage);
-        this.errorMessage = errorMessage;
-    }
-
-    /**
-     * Creates an error message to be displayed to the user.
-     *
-     * @return String containing the output message.
-     */
-    @Override
-    public String toString() {
-        return "OOPS!!! " + this.errorMessage;
-    }
-}
-
-class InvalidDateTime extends DukeException {
-    private final String errorMessage;
-    /**
-     * Initialises a new instance of InvalidDateTime.
-     *
-     * @param errorMessage The error message that is related to the exception.
-     */
-    public InvalidDateTime(String errorMessage) {
-        super(errorMessage);
-        this.errorMessage = errorMessage;
-    }
-
-    /**
-     * Creates an error message to be displayed to the user.
-     *
-     * @return String containing the output message.
-     */
-    @Override
-    public String toString() {
-        return "OOPS!!! " + this.errorMessage;
-    }
-}
