+++ src/main/java/duke/exception/InvalidCommand.java
@@ -1,2 +1,25 @@
-package PACKAGE_NAME;public class InvalidCommand {
+package duke.exception;
+
+public class InvalidCommand extends DukeException {
+
+    private final String errorMessage;
+    /**
+     * Initialises a new instance of duke.exception.InvalidCommand.
+     *
+     * @param errorMessage The error message that is related to the exception.
+     */
+    public InvalidCommand(String errorMessage) {
+        super(errorMessage);
+        this.errorMessage = errorMessage;
+    }
+
+    /**
+     * Creates an error message to be displayed to the user.
+     *
+     * @return String containing the output message.
+     */
+    @Override
+    public String toString() {
+        return "OOPS!!! " + this.errorMessage;
+    }
 }
