+++ src/main/java/duke/misc/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.misc;
 
 import duke.task.Deadline;
 import duke.task.Event;
