+++ src/main/java/duke/task/Event.java
@@ -1,6 +1,4 @@
-package duke;
-
-import duke.task.Task;
+package duke.task;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
